feat: expense section

+++ lib/core/theme/utils/format_currency.dart
@@ -1,6 +1,6 @@
 import 'package:intl/intl.dart';
 
-String formatCurrency(int? valueInCents) {
+String formatCurrency(double? valueInCents) {
   if (valueInCents == null) return '';
   return NumberFormat.currency(
     locale: 'pt_BR',
